requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langgraph>=0.0.30
langchain-groq>=0.1.0
firecrawl-py>=0.0.10
streamlit>=1.33.0
pydantic>=2.6.0
python-dotenv>=1.0.1
